+++ Client/Frontend/Login Management/LoginDataSource.swift
@@ -43,10 +43,6 @@ class LoginDataSource: NSObject, UITableViewDataSource {
     @objc func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = LoginListTableViewCell(style: .subtitle, reuseIdentifier: CellReuseIdentifier)
 
-        // Need to override the default background multi-select color to support theming
-        cell.multipleSelectionBackgroundView = UIView()
-        cell.applyTheme()
-
         if indexPath.section == LoginsSettingsSection {
             let hideSettings = viewModel.searchController?.isActive ?? false || tableView.isEditing
             let setting = indexPath.row == 0 ? boolSettings.0 : boolSettings.1
@@ -81,43 +77,3 @@ class LoginDataSource: NSObject, UITableViewDataSource {
     }
 }
 
-class LoginListTableViewCell: ThemedTableViewCell {
-    lazy var breachAlertImageView: UIImageView = {
-        let image = UIImage(named: "Breached Website")
-        let imageView = UIImageView(image: image)
-        imageView.isHidden = true
-        return imageView
-    }()
-    let breachAlertSize: CGFloat = 24
-
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        accessoryType = .disclosureIndicator
-        contentView.addSubview(breachAlertImageView)
-        breachAlertImageView.snp.remakeConstraints { make in
-            make.centerY.equalTo(contentView)
-            make.trailing.equalTo(contentView.snp.trailing).offset(-LoginTableViewCellUX.HorizontalMargin)
-            make.width.equalTo(breachAlertSize)
-            make.height.equalTo(breachAlertSize)
-        }
-
-        textLabel?.snp.remakeConstraints({ make in
-            make.leading.equalTo(contentView).offset(LoginTableViewCellUX.HorizontalMargin)
-            make.trailing.equalTo(breachAlertImageView.snp.leading).offset(-LoginTableViewCellUX.HorizontalMargin/2)
-            make.top.bottom.equalTo(contentView)
-            make.centerY.equalTo(contentView)
-            if let detailTextLabel = self.detailTextLabel {
-                make.bottom.equalTo(detailTextLabel.snp.top)
-                make.top.equalTo(contentView.snp.top).offset(LoginTableViewCellUX.HorizontalMargin)
-            }
-        })
-
-        // Need to override the default background multi-select color to support theming
-        self.multipleSelectionBackgroundView = UIView()
-        self.applyTheme()
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
